execution was splitted into more docments, separating, the ast with basic commands, eeuction for builtin, and there redirections with fds
