+++ out
@@ -1,0 +1,10 @@
+gnl
+includes
+libft
+Makefile
+minishell
+objs
+out
+README.rd
+rl.supp
+sources
