+++ sources/[4]execution/[3]traversing_ast.c
@@ -6,7 +6,7 @@
 /*   By: motero <motero@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/26 18:21:25 by motero            #+#    #+#             */
-/*   Updated: 2023/02/02 21:51:36 by motero           ###   ########.fr       */
+/*   Updated: 2023/02/02 22:06:53 by motero           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,26 +41,7 @@ void	simple_command_traverse(t_minishell *msh, t_ast *root, int *i)
 
 	if (root == NULL)
 		return ;
-	if (msh->fd_out == -1 || msh->fd_in == -1)
-		error_safe_exit("FD ERROR\n");
-	if (msh->fd_in > 0)
-	{
-		dup2(msh->fd_in, STDIN_FILENO);
-		close(msh->fd_in);
-	}
-	if (msh->fd_out > 0)
-	{
-		dup2(msh->fd_out, STDOUT_FILENO);
-		close(msh->fd_out);
-	}
-	else if (*i > 0)
-		dup2(root->pipe_fd[1], STDOUT_FILENO);
-	close(root->pipe_fd[1]);
-	if (msh->fd_in > 0 && *i > 0)
-	{
-		dup2(msh->fd_in, STDIN_FILENO);
-		close(msh->fd_in);
-	}
+	fds_handlings(msh, root, i);
 	setup_sig_child();
 	left = root->left;
 	if (!left)
@@ -89,66 +70,7 @@ void	argument_traverse(t_minishell *msh, t_ast *root, int *i)
 		return ;
 	left = root->left;
 	right = root->right;
-	//printf("Start of a ARGUMENT\n");
 	main_execution(msh, left, i);
 	main_execution(msh, right, i);
-	(void)i;
-}
-
-//ast root a redirection
-void	redirection_traverse(t_minishell *msh, t_ast *root, int *i)
-{
-	t_ast	*left;
-	t_ast	*right;
-
-	if (root == NULL)
-		return ;
-	left = root->left;
-	right = root->right;
-	(void) right;
-	printf("Start of a REDIRECTION\n");
-	if (ft_strcmp(">", left->data) == 1)
-		redirect_output(msh, root, i);
-	else if (ft_strcmp("<", left->data) == 1)
-		redirect_input(msh, root, i);
-	else if (ft_strcmp(">>", left->data) == 1)
-		redirect_append(msh, root, i);
-	else if (ft_strcmp("<<", left->data) == 1)
-		redirect_heredoc(msh, root, i);
-	printf("\n");
-	(void)i;
-}
-
-//ast root a simple builtin
-void	simple_builtin_traverse(t_minishell *msh, t_ast *root, int *i)
-{
-	t_ast	*left;
-	t_ast	*right;
-
-	if (root == NULL)
-		return ;
-	left = root->left;
-	right = root->right;
-	main_execution(msh, left, i);
-	main_execution(msh, right, i);
-	printf("Start of a SIMPLE_BUILTIN\n");
-	printf("Command: %s", left->data);
-	printf("Args: %s\n\n", right->data);
-	(void)i;
-}
-
-//ast root a complex builtin
-void	complex_builtin_traverse(t_minishell *msh, t_ast *root, int *i)
-{
-	t_ast	*left;
-	t_ast	*right;
-
-	if (root == NULL)
-		return ;
-	left = root->left;
-	right = root->right;
-	main_execution(msh, left, i);
-	main_execution(msh, right, i);
-	printf("Start of a CMPLX_BUILT\n");
 	(void)i;
 }
