+++ sources/[4]execution/[5]builtins_exec.c
@@ -1,0 +1,47 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   [5]builtins_exec.c                                 :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: motero <motero@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/02/02 22:06:23 by motero            #+#    #+#             */
+/*   Updated: 2023/02/02 22:06:46 by motero           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "execution.h"
+
+//ast root a simple builtin
+void	simple_builtin_traverse(t_minishell *msh, t_ast *root, int *i)
+{
+	t_ast	*left;
+	t_ast	*right;
+
+	if (root == NULL)
+		return ;
+	left = root->left;
+	right = root->right;
+	main_execution(msh, left, i);
+	main_execution(msh, right, i);
+	printf("Start of a SIMPLE_BUILTIN\n");
+	printf("Command: %s", left->data);
+	printf("Args: %s\n\n", right->data);
+	(void)i;
+}
+
+//ast root a complex builtin
+void	complex_builtin_traverse(t_minishell *msh, t_ast *root, int *i)
+{
+	t_ast	*left;
+	t_ast	*right;
+
+	if (root == NULL)
+		return ;
+	left = root->left;
+	right = root->right;
+	main_execution(msh, left, i);
+	main_execution(msh, right, i);
+	printf("Start of a CMPLX_BUILT\n");
+	(void)i;
+}
